use shared spring app and test support lib from jcenter
