+++ gradle/dependencies.gradle
@@ -1,11 +1,7 @@
 dependencies {
-    compile "org.springframework.boot:spring-boot-starter-web:$libSpringBootVersion"
-    compile "org.springframework.boot:spring-boot-starter-actuator:$libSpringBootVersion"
+    compile 'com.github.mrduguo.spring:spring-app:0.3.0-180320-160539-532363b-7'
     compile "org.codehaus.groovy:groovy-all:$libGroovyVersion"
-    compile 'io.springfox:springfox-swagger2:2.8.0'
-    compile 'io.springfox:springfox-swagger-ui:2.8.0'
 
-    testCompile "org.springframework.boot:spring-boot-starter-test:$libSpringBootVersion"
+    testCompile 'com.github.mrduguo.spring:spring-test:0.3.0-180320-161244-f62665d-6'
     testCompile "org.springframework.boot:spring-boot-starter-aop:$libSpringBootVersion"
-    testCompile 'org.spockframework:spock-spring:1.0-groovy-2.4'
 }
