+++ src/main/groovy/com/github/h2020_5gtango/vnv/tee/config/BearerAuthorizationInterceptor.groovy
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2002-2016 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.github.h2020_5gtango.vnv.tee.config
+
+import com.github.h2020_5gtango.vnv.tee.restclient.UserSessionManager
+import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.http.HttpRequest
+import org.springframework.http.client.ClientHttpRequestExecution
+import org.springframework.http.client.ClientHttpRequestInterceptor
+import org.springframework.http.client.ClientHttpResponse
+import org.springframework.stereotype.Component
+
+@Component
+class BearerAuthorizationInterceptor implements ClientHttpRequestInterceptor {
+
+    @Autowired
+    UserSessionManager userSessionManager
+
+
+    @Override
+    ClientHttpResponse intercept(HttpRequest request, byte[] body,
+                                        ClientHttpRequestExecution execution) throws IOException {
+        request.getHeaders().add('Authorization', 'bearer ' + userSessionManager.retrieveValidBearerToken())
+        execution.execute(request, body)
+    }
+
+}
