+++ src/main/groovy/com/github/h2020_5gtango/vnv/tee/model/NetworkServiceInstance.groovy
@@ -1,0 +1,12 @@
+package com.github.h2020_5gtango.vnv.tee.model
+
+
+class NetworkServiceInstance {
+
+    String networkServiceInstanceId
+
+    String networkServiceId
+
+    Map runtime
+
+}
