--- src/main/groovy/eu/h2020_5gtango/vnv/tee/model/NetworkServiceInstance.groovy
@@ -1,12 +1,0 @@
-package eu.h2020_5gtango.vnv.tee.model
-
-
-class NetworkServiceInstance {
-
-    String networkServiceInstanceId
-
-    String networkServiceId
-
-    Map runtime
-
-}
