--- src/main/groovy/eu/h2020_5gtango/vnv/tee/model/TestSuiteResult.groovy
@@ -1,19 +1,0 @@
-package eu.h2020_5gtango.vnv.tee.model
-
-import groovy.transform.EqualsAndHashCode
-
-import java.security.MessageDigest
-
-@EqualsAndHashCode
-class TestSuiteResult {
-
-    String testSuiteResultId
-
-    String testPlanId
-    String networkServiceInstanceId
-    String testSuiteId
-
-    String status
-
-    Map details
-}
